fix: move utils functions to sign.service.ts file
